pyproject.toml
[project]
name = "heart-disease-risk-prediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.7",
    "lightgbm>=4.5.0",
    "matplotlib>=3.9.2",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "ruff>=0.7.4",
    "scikit-learn>=1.5.2",
    "seaborn>=0.13.2",
    "ucimlrepo>=0.0.7",
]
